Fix high-priority bugs: Remove duplicate filters and fix year calculations
- Remove duplicate Rank Math OpenGraph filter registrations (lines 24-35)
- Remove duplicate Elementor/CRP filters from seopress-support.php (already in others.php)
- Fix incorrect esc_html() usage in year shortcodes [nyear], [nnyear], [pyear], [ppyear]
  - Now properly cast date_i18n('Y') to int before arithmetic operations
  - Prevents string concatenation instead of mathematical addition/subtraction

These changes improve performance by eliminating redundant filter processing and ensure correct year calculations.

+++ includes/shortcodes.php
@@ -129,25 +129,25 @@ add_shortcode( 'pmonthyear', 'dmyip_pmonthyear' );
 // [nyear] shortcode.
 add_shortcode( 'nyear', 'dmyip_rmd_next_year' );
 function dmyip_rmd_next_year() {
-    return esc_html( date_i18n( 'Y' ) + 1 );
+    return esc_html( (int) date_i18n( 'Y' ) + 1 );
 }
 
 // [nnyear] shortcode.
 add_shortcode( 'nnyear', 'dmyip_rmd_next_next_year' );
 function dmyip_rmd_next_next_year() {
-    return esc_html( date_i18n( 'Y' ) + 2 );
+    return esc_html( (int) date_i18n( 'Y' ) + 2 );
 }
 
 // [pyear] shortcode.
 add_shortcode( 'pyear', 'dmyip_rmd_previous_year' );
 function dmyip_rmd_previous_year() {
-    return esc_html( date_i18n( 'Y' ) - 1 );
+    return esc_html( (int) date_i18n( 'Y' ) - 1 );
 }
 
 // [ppyear] shortcode.
 add_shortcode( 'ppyear', 'dmyip_rmd_previous_previous_year' );
 function dmyip_rmd_previous_previous_year() {
-    return esc_html( date_i18n( 'Y' ) - 2 );
+    return esc_html( (int) date_i18n( 'Y' ) - 2 );
 }
 
 // Day shortcodes.
